+++ atmos/lib_vulkan/vulkan/texture/vulkan_texture2d.cpp
@@ -135,7 +135,7 @@ vulkan_texture2d::vulkan_texture2d(
 	, _gl_handle(gl_handle)
 	, _gpu_mem_metric(perf::category::MEM_USAGE, "GPU Texture 2D Mem")
 {
-	_gpu_mem_metric.add((float)size);
+	_gpu_mem_metric.add(size);
 }
 
 vulkan_texture2d::~vulkan_texture2d()
